+++ open-sse/providers/registry/index.js
@@ -121,6 +121,10 @@ import p118 from "./selfhosted-tts.js";
 import p119 from "./selfhosted-embedding.js";
 import p120 from "./fish-audio.js";
 import p121 from "./alitp-intl.js";
+import p122 from "./longcat.js";
+import p123 from "./sensenova.js";
+import p124 from "./dots.js";
+import p125 from "./bai.js";
 
 export default [
   p0,
@@ -243,4 +247,8 @@ export default [
   p119,
   p120,
   p121,
+  p122,
+  p123,
+  p124,
+  p125,
 ];